llamaindex_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
llama-index-core==0.1.0
llama-index-workflow==0.1.0
llama-index-llms-openai==0.1.0
python-multipart==0.0.6
httpx==0.25.2

